+++ Example/app/App.js
@@ -273,6 +273,7 @@ export default class App extends Component {
           action: Actions.INBOX_NOTIFICATION_CLICKED,
           name: 'pushInboxNotificationClickedEvent',
         },
+        { action: Actions.FETCH_INBOX, name: 'fetchInbox' },
       ],
     },
     {
@@ -348,6 +349,7 @@ export default class App extends Component {
       subCategory: [
         { action: Actions.DISPLAY_UNIT_ID, name: 'getUnitID' },
         { action: Actions.ALL_DISPLAY_UNITS, name: 'getAllDisplayUnits' },
+        { action: Actions.PUSH_DISPLAY_UNIT_ELEMENT_CLICKED, name: 'pushDisplayUnitElementClickedEventForID' },
       ],
     },
     {
@@ -513,6 +515,9 @@ export default class App extends Component {
       case Actions.INBOX_NOTIFICATION_CLICKED:
         AppUtils.pushInboxNotificationClicked();
         break;
+      case Actions.FETCH_INBOX:
+        AppUtils.fetchInbox();
+        break;
       case Actions.PUSH_EVENT:
         AppUtils.pushevent();
         break;
@@ -578,6 +583,9 @@ export default class App extends Component {
         break;
       case Actions.ALL_DISPLAY_UNITS:
         AppUtils.getAllDisplayUnits();
+        break;
+      case Actions.PUSH_DISPLAY_UNIT_ELEMENT_CLICKED:
+        AppUtils.pushDisplayUnitElementClicked();
         break;
       case Actions.PRODUCT_CONFIG_FETCH:
         AppUtils.fetch();